Start the game in GAME_TRANSITION instead of GAME_ACTIVE

+++ src/font.cpp
@@ -58,7 +58,7 @@ Font::Font() {
             texture,
             glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
             glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-            face->glyph->advance.x};
+            static_cast<unsigned int>(face->glyph->advance.x)};
         this->Characters.insert(std::pair<char, Character>(c, character));
     }
     glBindTexture(GL_TEXTURE_2D, 0);
